// typings.d.ts
export type ExploreData = {
  img: string;
  location: string;
  distance: string;
};

export type CardsData = {
  img: string;
  title: string;
};

export type SearchResultsData = {
  img: string;
  title: string;
  location: string;
  description: string;
  star: number;
  price: string;
  total: string;
  long: number;
  lat: number;
};